+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 ### Changed
 - WMS-Time: Shrunk wms-time slider for mobile views, adjusted Position on desktop.
 - getFeatureInfo: gfi now works on BaseLayers as well as SubjectLayers.
+- Menutitle: Allow html in title text.
 
 ### Deprecated